Apply handler for screen from CvEventManager
